hwmon: lm95234: Fix type incompatibility with non-macro find_closest
The update_intervals array was previously declared as u16 but used with
find_closest(). With find_closest() now implemented as a function taking
signed int parameters instead of a macro, passing unsigned arrays causes
type incompatibility errors. This patch changes the array type from
u16 to int to ensure compatibility with the function signature and
prevent compilation errors.

+++ drivers/hwmon/lm95234.c
@@ -240,7 +240,7 @@ static int lm95234_temp_read(struct device *dev, u32 attr, int channel, long *va
 	return 0;
 }
 
-static u16 update_intervals[] = { 143, 364, 1000, 2500 };
+static int update_intervals[] = { 143, 364, 1000, 2500 };
 
 static int lm95234_chip_write(struct device *dev, u32 attr, long val)
 {
